+++ website/src/views/playground.ts
@@ -34,7 +34,7 @@ const TETRIS_EXAMPLE = await FS.readFile(
 );
 
 const XSTATE_EXAMPLE = await FS.readFile(
-	Path.join(__dirname, "../../examples/xstate-calculator.ts"),
+	Path.join(__dirname, "../../examples/xstate-calculator.tsx"),
 	"utf8",
 );
